requirements.txt
crewai>=0.28.8
crewai-tools>=0.1.6
langchain-google-genai>=1.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pyyaml>=6.0
